made code more standards compliant, generally cleaned up, and resolved #150

+++ src/qvinfodialog.cpp
@@ -39,7 +39,8 @@ void QVInfoDialog::updateInfo()
     QLocale locale = QLocale::system();
     QMimeDatabase mimedb;
     QMimeType mime = mimedb.mimeTypeForFile(selectedFileInfo.filePath(), QMimeDatabase::MatchContent);
-    const float megapixels = float(int((static_cast<float>((width*height)))/1000000 * 10 + 0.5f)) / 10 ;
+    //this is just math to figure the megapixels and then round it to the tenths place
+    const double megapixels = static_cast<double>(static_cast<int>((static_cast<double>((width*height)))/1000000 * 10 + 0.5)) / 10 ;
 
     ui->nameLabel->setText(selectedFileInfo.fileName());
     ui->typeLabel->setText(mime.name());
